Facebook integration demo
Allows user to login to their Facebook account and if the login is
successful then it shows a list of friends (e.g. taggable friends.) If
an item in the list is clicked a toast message pops out, it should show
the friend’s birthday. However, it cannot access this information, even
though the permission to show friend’s birthday is set.